requirements.txt
python-telegram-bot==13.7
requests==2.26.0
Flask==2.1.0
SQLAlchemy==1.4.23
python-dotenv==0.19.0
# Agrega más bibliotecas según sea necesario